+++ Realm/RLMArrayTableView.mm
@@ -206,8 +206,7 @@ static inline void RLMArrayTableViewValidateInWriteTransaction(RLMArrayTableView
 {
     RLMArrayTableViewValidate(self);
 
-    tightdb::Query query = _backingView.get_parent().where();
-    query.tableview(_backingView);
+    tightdb::Query query = _backingView.get_parent().where(&(_backingView));
     
     // apply order
     RLMArrayTableView *ar = [RLMArrayTableView arrayWithObjectClassName:self.objectClassName